WIP: Don't block sends with unfinished x-chain tfr.
Created separate modal for x-chain transfer.

Removed blocking from send modal and base send modal no longer handles x-chain
functionality. Similarly the x-chain modal does not accept same-chain transfers(??).
